Refactor RabitMq TransferLogDataBase

+++ MicroRabbit.Transfer.Domain/Models/TransferLog.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int FromAccount { get; set; }
-        public decimal ToAccount { get; set; }
+        public int ToAccount { get; set; }
         public decimal Amount { get; set; }
     }
 }
